Restructure how dependency injection is used
The original way passed an IServiceCollection as a parameter in the App constructor. Due to the AppBuilder constructor that accepts an Application factory as parameter is not exposed on Android and iOS, it was hacky on Android, and impossible on iOS to create the AppBuilder. The IServiceCollection now lives in a static member of a separate class.