Do not send cookies with requests
The /auth/login endpoint sets a csrf_access_token
cookie, which was sent along subsequent requests on
a TLS connection (only on TLS because of the Secure
flag). This is unnecessary when sending the JWT in
a header, but made the requests fail because the
CSRF header was not present.