Separate session data push/pull into individual functions
This change also ensures that only sessions that has no data field on
the server side are pushed. Moreover, we try to push data for all of
these sessions, not just the ones that were synced now (this way we
can somewhat mitigate past sync errors). The same mechanism works
during pulls too: we try to pull all session data we do not have
locally.